+++ react/FetchMore.js
@@ -1,23 +1,24 @@
 import React from 'react'
+// eslint-disable-next-line no-restricted-imports
 import { path } from 'ramda'
 import classNames from 'classnames'
 import PropTypes from 'prop-types'
+import { useSearchPage } from 'vtex.search-page-context/SearchPageContext'
+
 import FetchMoreButton from './components/loaders/FetchMoreButton'
 import LoadingSpinner from './components/loaders/LoadingSpinner'
 import { PAGINATION_TYPE } from './constants/paginationType'
 import { useFetchMore } from './hooks/useFetchMore'
-
-import { useSearchPage } from 'vtex.search-page-context/SearchPageContext'
-
 import styles from './searchResult.css'
 
-const FetchMore = ({ htmlElementForButton = 'button '}) => {
+const FetchMore = ({ htmlElementForButton = 'button' }) => {
   const { pagination, searchQuery, maxItemsPerPage, page } = useSearchPage()
   const products = path(['data', 'productSearch', 'products'], searchQuery)
   const recordsFiltered = path(
     ['data', 'productSearch', 'recordsFiltered'],
     searchQuery
   )
+
   const fetchMore = path(['fetchMore'], searchQuery)
   const queryData = {
     query: path(['variables', 'query'], searchQuery),
@@ -26,7 +27,7 @@ const FetchMore = ({ htmlElementForButton = 'button '}) => {
     priceRange: path(['variables', 'priceRange'], searchQuery),
   }
 
-  const { handleFetchMoreNext, loading, to } = useFetchMore({
+  const { handleFetchMoreNext, loading, to, nextPage } = useFetchMore({
     page,
     recordsFiltered,
     maxItemsPerPage,
@@ -53,6 +54,7 @@ const FetchMore = ({ htmlElementForButton = 'button '}) => {
           loading={loading}
           showProductsCount={false}
           htmlElementForButton={htmlElementForButton}
+          nextPage={nextPage}
         />
       </div>
     )
@@ -64,6 +66,10 @@ const FetchMore = ({ htmlElementForButton = 'button '}) => {
 FetchMore.propTypes = {
   /* html element to render for fetch more button */
   htmlElementForButton: PropTypes.string,
+}
+
+FetchMore.schema = {
+  title: 'admin/editor.search-result.fetch-more',
 }
 
 export default FetchMore