+++ react/GalleryLegacy.tsx
@@ -16,6 +16,9 @@ import GalleryRow from './components/GalleryRow'
 import ProductListEventCaller from './utils/ProductListEventCaller'
 import SettingsContext from './components/SettingsContext'
 import type { Product } from './Gallery'
+import type { PreferredSKU } from './GalleryLayout'
+import { useBreadcrumb } from './hooks/useBreadcrumb'
+import { useSearchTitle } from './hooks/useSearchTitle'
 
 /** Layout with one column */
 const ONE_COLUMN_LAYOUT = 1
@@ -39,6 +42,8 @@ export interface GalleryProps {
   lazyItemsRemaining: number
   customSummaryInterval?: number
   CustomSummary?: ComponentType
+  /** Logic to enable which SKU will be the selected item */
+  preferredSKU?: PreferredSKU
 }
 
 /**
@@ -59,12 +64,19 @@ const Gallery: React.FC<GalleryProps> = ({
   lazyItemsRemaining,
   customSummaryInterval,
   CustomSummary,
+  preferredSKU,
 }) => {
   const { isMobile } = useDevice()
-  const { trackingId = 'Search result' } = useContext(SettingsContext) || {}
+  const { trackingId } = useContext(SettingsContext) || {}
   const handles = useCssHandles(CSS_HANDLES)
   const { getSettings } = useRuntime()
   const responsiveMaxItemsPerRow = useResponsiveValue(maxItemsPerRow)
+  const breadcrumb = useBreadcrumb()
+  const searchTitle = useSearchTitle(breadcrumb ?? []).trim()
+
+  // Not using ?? operator because trackingId and searchTitle can be ''
+  // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
+  const listName = trackingId || searchTitle || 'Search result'
 
   const layoutMode = isMobile ? mobileLayoutMode : 'normal'
 
@@ -109,7 +121,7 @@ const Gallery: React.FC<GalleryProps> = ({
     ?.enableSearchRenderingOptimization
 
   return (
-    <ProductListProvider listName={trackingId}>
+    <ProductListProvider listName={listName}>
       <div className={galleryClasses}>
         {rows.map((rowProducts, index) => (
           <GalleryRow
@@ -120,8 +132,10 @@ const Gallery: React.FC<GalleryProps> = ({
             displayMode={layoutMode}
             itemsPerRow={itemsPerRow}
             rowIndex={index}
+            listName={listName}
             customSummaryInterval={customSummaryInterval}
             CustomSummary={CustomSummary}
+            preferredSKU={preferredSKU}
           />
         ))}
         {typeof lazyItemsRemaining === 'number' && lazyItemsRemaining > 0 && (