+++ react/components/GalleryLayoutRow.tsx
@@ -1,10 +1,12 @@
-import React, { ComponentType, memo } from 'react'
-import { useRenderOnView } from '../hooks/useRenderOnView'
+import type { ComponentType } from 'react'
+import React, { memo } from 'react'
 import { useCssHandles, applyModifiers } from 'vtex.css-handles'
 import classNames from 'classnames'
 
+import { useRenderOnView } from '../hooks/useRenderOnView'
 import GalleryItem from './GalleryLayoutItem'
 import type { Product } from '../Gallery'
+import type { PreferredSKU } from '../GalleryLayout'
 
 const CSS_HANDLES = ['galleryItem'] as const
 
@@ -17,6 +19,9 @@ interface GalleryLayoutRowProps {
   products: Product[]
   summary: unknown
   rowIndex: number
+  listName: string
+  /** Logic to enable which SKU will be the selected item */
+  preferredSKU?: PreferredSKU
 }
 
 const GalleryLayoutRow: React.FC<GalleryLayoutRowProps> = ({
@@ -28,6 +33,8 @@ const GalleryLayoutRow: React.FC<GalleryLayoutRowProps> = ({
   summary,
   currentLayoutName,
   rowIndex,
+  listName,
+  preferredSKU
 }) => {
   const handles = useCssHandles(CSS_HANDLES)
 
@@ -48,9 +55,10 @@ const GalleryLayoutRow: React.FC<GalleryLayoutRowProps> = ({
   return (
     <>
       {products.map((product, index) => {
+        const absoluteProductIndex = rowIndex * itemsPerRow + index + 1
         return (
           <div
-            key={product.productId}
+            key={product.cacheId}
             style={style}
             className={classNames(
               applyModifiers(handles.galleryItem, [
@@ -65,7 +73,9 @@ const GalleryLayoutRow: React.FC<GalleryLayoutRowProps> = ({
               item={product}
               summary={summary}
               displayMode={displayMode}
-              position={(rowIndex * itemsPerRow) + index}
+              position={absoluteProductIndex}
+              listName={listName}
+              preferredSKU={preferredSKU}
             />
           </div>
         )